ai_toolwear_ncs/requirements.txt
absl-py==2.1.0
contourpy==1.1.1
cycler==0.12.1
distro==1.6.0
eii-configmgr @ file:///app/eii_configmgr-4.0.2-cp38-cp38-manylinux2014_x86_64.whl
eii-msgbus @ file:///app/eii_msgbus-4.0.0-cp38-cp38-manylinux2014_x86_64.whl
filelock==3.16.1
fonttools==4.53.0
fsspec==2024.9.0
h5py==3.11.0
imageio==2.35.1
importlib_resources==6.4.0
imutils==0.5.4
Jinja2==3.1.4
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.0.3
Pillow==9.5.0
protobuf==3.20.3
pycairo==1.16.2
PyGObject==3.36.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
scikit-build==0.12.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
tflite-runtime==2.14.0
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
